Atualização de dos Services

+++ src/main/java/br/com/eureka/eureka/repository/FonteRecursoRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FonteRecursoRepository extends JpaRepository<FonteRecurso, Integer> {
+public interface FonteRecursoRepository extends JpaRepository<FonteRecurso, Long> {
+    // Métodos personalizados do repositório podem ser adicionados aqui, se necessário
 }
