+++ src/main/java/br/com/eureka/eureka/repository/GrupoDespesaRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GrupoDespesaRepository extends JpaRepository<GrupoDespesa, Integer> {
+public interface GrupoDespesaRepository extends JpaRepository<GrupoDespesa, Long> {
+    // Aqui é possível adicionar métodos personalizados do repositório, se necessário
 }
